@@ -3,11 +3,10 @@ package com.dandy.msavaliadorcredito;
 import org.springframework.amqp.rabbit.annotation.EnableRabbit;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication
-@EnableEurekaClient
+// @EnableEurekaClient - A NOVA VERSÃO DO SPRING CLOUD NÃO TEM ESSA ANOTAÇÃO, BASTA TER A DEPENDENCIA E A CONFIG NA PROPERTY
 @EnableFeignClients
 @EnableRabbit
 public class MsavaliadorcreditoApplication {
